+++ bin/blocks/cta.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+# @Author: Roni Laukkarinen
+# @Date:   2021-11-23 18:23:24
+# @Last Modified by:   Roni Laukkarinen
+# @Last Modified time: 2021-11-23 18:24:33
+elif [ "$BLOCK_NAME" == "cta" ]; then
+  echo 'This is a cta block';