+++ bin/blocks/hero.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+# @Author: Roni Laukkarinen
+# @Date:   2021-11-23 18:23:24
+# @Last Modified by:   Roni Laukkarinen
+# @Last Modified time: 2021-11-23 18:24:07
+if [ "$BLOCK_NAME" == "hero" ]; then
+  echo 'This is a hero block';