+++ bin/tasks/replaces.sh
@@ -1,0 +1,15 @@
+#!/bin/bash
+# @Author: Roni Laukkarinen
+# @Date:   2021-11-23 18:26:55
+# @Last Modified by:   Roni Laukkarinen
+# @Last Modified time: 2021-11-23 18:28:35
+echo ""
+echo "${BOLDGREEN}All done!${TXTRESET} Block created: ${BLOCK_NAME}"
+echo ""
+echo "Front end: https://${PROJECT_NAME}.test"
+echo "WP-admin:  https://${PROJECT_NAME}.test/wp/wp-login.php"
+echo ""
+echo "Block generated. Go to https://$PROJECT_NAME.test and start coding your block.{TXTRESET}"
+echo ""
+echo "Thanks for using newblock start script ${SCRIPT_LABEL}, v${SCRIPT_VERSION}!"
+echo ""