+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### 1.0.1
 
+* First working draft of the `newblock` script
+* Fix block name tag class and transition
 * Add Inter font weights
 * Add block amount counter with real time filtering